Removed line and polygon from the spatial coverage section
When dealing with resource level metadata, a point location or a bounding box is sufficient to depict the spatial extent of the resource content. However, when it comes to file level metadata, we may want to consider more specific options such as point, line, polygon, and bounding box to accurately describe the spatial coverage of each dataset.